+++ actix-session/CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Replace the usages of `HashMap` with `serde::Map`.
+- Fix double serialization when inserting values into the session.
+
 ## 0.8.0
 
 - Set secure attribute when adding a session removal cookie.